move display_options to user_interface module
